Refactoring - introduce poetry - load environment variables from .env file - move source code into subfolder
