+++ env.example
@@ -1,0 +1,5 @@
+HUGGING_FACE_ACCESS_TOKEN=<your_access_token> #see TL;DR of https://huggingface.co/pyannote/speaker-diarization
+LLM_URL=<your_llm_url> #e.g. "https://api.openai.com/v1/chat/completions" for openAI
+LLM_MODEL=<your_llm_model> #e.g. "gpt-3.5-turbo"
+LLM_MAX_TOKENS=<yor_llm_max_tokens> #e.g. "4000"
+LLM_AUTHORIZATION=<your_llm_authorization> #e.g. "Bearer <Your OpenAI API key>"
